+++ spec/rubocop/cop/migration/change_table_references_spec.rb
@@ -4,6 +4,9 @@ RSpec.describe RuboCop::Cop::Migration::ChangeTableReferences, :config do
   it 'registers an offense when using references in a change_table' do
     expect_offense(<<~RUBY)
       change_table :subscriptions, bulk: true do |t|
+         t.string :name
+         t.string :email
+         t.boolean :admin, default: false
          t.references :user, null: false, foreign_key: true
          ^^^^^^^^^^^^ Use a combination of `t.bigint`, `t.index` and `t.foreign_key` in a change_table to add a reference.Or `t.remove_foreign_key`, `t.remove` to remove a reference.
       end